requirements.txt
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
entrypoints==0.4
Faker==37.4.0
fastapi==0.116.1
favicon==0.7.0
fonttools==4.58.5
fpdf2==2.8.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
htbuilder==0.9.0
httplib2==0.22.0
ics==0.7.2
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lxml==6.0.0
Markdown==3.8.2
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
more-itertools==10.7.0
narwhals==1.45.0
numpy==2.1.3
packaging==25.0
pandas==2.3.0
pillow==11.3.0
plotly==6.2.0
prometheus_client==0.22.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
pymdown-extensions==10.16
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
st-annotated-text==4.0.2
st-theme==1.2.3
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.47.1
streamlit==1.46.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-chat==0.1.1
streamlit-confetti==0.1.0
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
TatSu==5.13.1
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.27.0
validators==0.35.0
watchdog==6.0.0
weasyprint==65.1
webencodings==0.5.1
wheel==0.45.1
zopfli==0.2.3.post1
